вынесена логика для изменения даты для компонентов в отдельный класс
